build: Update compilation command and move queue structs to header
