+++ image_processing.h
@@ -10,6 +10,20 @@ typedef pixelT *ptrPixelT;
 typedef ptrPixelT *imgT;
 typedef struct
 {
+    int x, y;
+} coordt1;
+typedef struct coordT
+{
+    int x, y; // pixel coordinates
+    struct coordT *suivant;
+} coordt;
+
+typedef struct
+{
+    coordt *tete, *queue;
+} filet;
+typedef struct
+{
     int lg;
     int h;
     int r;
